Move common gorm code to separate package
